+++ packages/uniforms-bridge-graphql/src/index.ts
@@ -1,2 +1,1 @@
-import './register';
 export { default, default as GraphQLBridge } from './GraphQLBridge';