+++ packages/uniforms/src/ValidatedForm.tsx
@@ -83,14 +83,11 @@ export function Validated<Base extends typeof BaseForm>(Base: Base) {
 
       const { model, schema, validate, validator } = this.props;
       if (schema !== prevProps.schema || validator !== prevProps.validator) {
-        this.setState(
-          state => ({ validator: state.bridge.getValidator(validator) }),
-          () => {
-            if (shouldRevalidate(validate, this.state.validate)) {
-              this.onValidate();
-            }
-          },
-        );
+        this.setState({ validator: schema.getValidator(validator) }, () => {
+          if (shouldRevalidate(validate, this.state.validate)) {
+            this.onValidate();
+          }
+        });
       } else if (
         !isEqual(model, prevProps.model) &&
         shouldRevalidate(validate, this.state.validate)